requirements.txt
altair==5.2.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
branca==0.8.2
cachetools==5.5.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
distro==1.9.0
fastapi==0.128.0
folium==0.20.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==7.2.1
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.26.0
mdurl==0.1.2
numpy==1.26.4
openai==2.16.0
packaging==23.2
pandas==2.1.4
pillow==10.4.0
protobuf==4.25.8
pyarrow==14.0.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rich==13.9.4
rpds-py==0.30.0
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==3.2.0
starlette==0.50.0
streamlit==1.31.0
streamlit-folium==0.26.1
tenacity==8.5.0
toml==0.10.2
toolz==1.1.0
tornado==6.5.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
validators==0.35.0
xyzservices==2025.11.0
zipp==3.23.0
